Add DS support for slice_scatter
* Adds dynamic shape (DS) support for slice_scatter op
* Handle negative indices as parameters for DS
* Adds DS test case for slice_scatter

+++ habana_eager/graph_dynamic.cpp
@@ -22,7 +22,7 @@ namespace graph {
 int64_t GetSymintValue(torch::jit::Stack& original_stack, uint64_t index) {
   int64_t value;
   HABANA_ASSERT(original_stack[index].isScalar() == 1);
-  value = original_stack[index].toScalar().toInt();
+  value = original_stack[index].toScalar().toLong();
   return value;
 }
